+++ SimulideService/Migrations/CollabContextModelSnapshot.cs
@@ -41,8 +41,8 @@ namespace SimulideService.Migrations
                     b.Property<DateTime>("UpdatedAt")
                         .HasColumnType("timestamp with time zone");
 
-                    b.Property<long>("Version")
-                        .HasColumnType("bigint");
+                    b.Property<int>("Version")
+                        .HasColumnType("integer");
 
                     b.HasKey("Id");
 
@@ -64,17 +64,17 @@ namespace SimulideService.Migrations
                     b.Property<Guid>("DocumentId")
                         .HasColumnType("uuid");
 
-                    b.Property<long>("Length")
-                        .HasColumnType("bigint");
+                    b.Property<int>("Length")
+                        .HasColumnType("integer");
 
-                    b.Property<long>("Position")
-                        .HasColumnType("bigint");
+                    b.Property<int>("Position")
+                        .HasColumnType("integer");
 
                     b.Property<int>("Type")
                         .HasColumnType("integer");
 
-                    b.Property<long>("Version")
-                        .HasColumnType("bigint");
+                    b.Property<int>("Version")
+                        .HasColumnType("integer");
 
                     b.HasKey("Id");
